Move Travis and Github helpers to inside the main project
