+++ src/main/interfaces/GithubHelperException.groovy
@@ -1,4 +1,4 @@
-package services
+package main.exception
 
 public class GithubHelperException extends Exception {
     public GithubHelperException(String message) {
