+++ src/main/interfaces/TravisHelperException.groovy
@@ -1,4 +1,4 @@
-package services
+package main.exception
 
 public class TravisHelperException extends Exception {
     public TravisHelperException(String message) {
